requirements.txt
streamlit
structlog
pypdf2
pydub
gtts
alembic==1.15.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.5
asttokens==3.0.0
async-lru==2.0.5
attrs==24.2.0
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.4.1
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorgram.py==1.2.0
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.4
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.14
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
Flask-CKEditor==1.0.0
Flask-Gravatar==0.5.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.1
fqdn==1.5.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
gspread==6.1.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.7
imageio==2.37.0
incremental==24.7.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==5.13.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.3.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
narwhals==1.42.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.1.0
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.1.2
prettytable==3.11.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
Protego==0.4.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pylint==3.3.1
pyOpenSSL==25.1.0
pyparsing==3.2.0
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.8.0
ratelim==0.1.6
redis==5.2.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spotipy==2.24.0
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.1
tinycss2==1.4.0
tldextract==5.3.0
tomlkit==0.13.2
tornado==6.5.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==25.5.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.35.0
w3lib==2.3.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wsproto==1.2.0
WTForms==3.2.1
zope.interface==7.2
